debug

+++ main.py
@@ -23,7 +23,7 @@ def main():
     person_boxes = outputs["instances"].pred_boxes[outputs["instances"].pred_classes == 0]
     person_images = []
     for box in person_boxes:
-        box = box.cpu().numpy()
+        box = box.cpu().numpy().astype(int)
         print(box)
         image = im[box[0]:box[2], box[1]: box[3]]
         person_images.append(image)
